test(pwa): enforce rendered icon in visual evidence
